+++ lib/nxt_pipeline.rb
@@ -12,4 +12,17 @@ module NxtPipeline
   class << self
     delegate :new, :call, to: Pipeline
   end
+
+  def configure(name, &block)
+    @configurations ||= {}
+    raise ArgumentError, "Configuration already defined for #{name}" if @configurations[name].present?
+
+    @configurations[name] = block
+  end
+
+  def configuration(name)
+    @configurations.fetch(name)
+  end
+
+  module_function :configure, :configuration
 end